+++ samewords/cli.py
@@ -1,30 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""Samewords annotates potentially ambiguous words in critical text editions
-made with LaTeX and reledmac.
-
-Usage: samewords [options] <file>
-
-Arguments:
-  <file>                Location of local file to be processed.
-
-Options:
-  --output <location>      Location of output. You can specify a filename as
-                           part of the address. If you don't do that, the name
-                           of the input file will be used.
-  --config-file <file>     JSON-formatted file with application configuration
-                           variables. For recognized keys and formatting of
-                           values, see the documentation on
-                           https://github.com/stenskjaer/samewords.
-                           [default: ~/.samewords.json]
-  -v, --version            Show version and exit.
-  -h, --help               Show this help message and exit.
+"""Command line interface director for the samewords script.
 """
 
 import json
 
 import samewords
-import docopt
+import argparse
 import os
 
 from samewords import settings
@@ -57,12 +39,41 @@ def parse_config_file(filename: str) -> object:
         return settings
     return None
 
+def parse_arguments():
+    parser = argparse.ArgumentParser(
+        prog='samewords',
+        usage='%(prog)s [options] FILE',
+        description='Annotate potentially ambiguous words in critical text '
+                    'editions made with LaTeX and reledmac.')
+    parser.add_argument('file', metavar='FILE', type=str, nargs=1,
+                        help='Location of local file to be processed.')
+    parser.add_argument(
+        '--output', dest='location', action='store',
+        help=("Location of the output. You can specify a filename as part of "
+              "the address. If you don't do that, the name of the input file "
+              "will be put in the directory specified. (default: '%("
+              "default)s')")
+    )
+    parser.add_argument(
+        '--config-file', dest='config', action='store',
+        help=("Location of the configuration file. For information on what it "
+             "may contain, see the README.")
+        )
+    parser.add_argument('-v', '--version', action='version',
+                        version='%(prog)s {}'.format(samewords.__version__),
+                        help="Show version and exit.")
+    return vars(parser.parse_args())
 
 def main():
-    args = docopt.docopt(__doc__, version=samewords.__version__)
-    filename = args['<file>']
-    output = args['--output']
-    parse_config_file(args['--config-file'])
+    # Read command line arguments
+    args = parse_arguments()
+
+    filename = args['file'][0]
+    output = args['location']
+    config = args['config']
+
+    if config:
+        parse_config_file(config)
 
     if not output:
         print(samewords.core.process_document(filename))
@@ -71,8 +82,8 @@ def main():
             _, output_filename = os.path.split(filename)
             output_dir = output
         elif os.path.isfile(output):
-            answer = input(
-                'The file {} already exists. Overwrite (y/n)? '.format(output))
+            answer = input('The file {} already exists. Overwrite (y/n)?'
+                           .format(output))
             if answer.lower() == 'y':
                 output_dir, output_filename = os.path.split(output)
             else: