+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Added
 - Proper handling of escaped LaTeX characters.
 
+### Removed
+- Dependecy on Docopt.
+
 ## [0.2.0] - 2018-01-27
 ### Added
 - Context length matching *exactly* 30 content words rather than *at least*.